+++ src/core/config/config.interfaces.ts
@@ -103,6 +103,8 @@ export interface HomebridgeUiConfig {
     cert?: string
     pfx?: string
     passphrase?: string
+    selfSigned?: boolean
+    selfSignedHostnames?: string[]
   }
   accessoryControl?: {
     debug?: boolean