+++ CHANGELOG.md
@@ -11,6 +11,7 @@ All notable changes to `homebridge-config-ui-x` will be documented in this file.
 - feat: add separate beta preferences for plugins (#2687) (@seidnerj)
 - further tweaks for node + plugin update preferences
 - open accessory info on tile loading icon click
+- first step of improvements to https settings (@donavanbecker)
 
 ### Homebridge Dependencies